refactor(home page): refactor the structure of the cards components
refactor the structure of the cards componenets

fix #12
