+++ src/components/Cards/__test__/TicketCard.test.js
@@ -1,5 +1,6 @@
 import renderer from "react-test-renderer";
-import TicketCard from "@/components/Home/TicketCard";
+
+import TicketCard from "@/components/Cards/TicketCard";
 
 it("renders correctly", () => {
     const tree = renderer.create(<TicketCard />).toJSON();
